+++ geom/slippy/tile.go
@@ -2,7 +2,7 @@ package slippy
 
 import "math"
 
-func NewTile(z, x, y uint64, buffer float64, srid uint64) *Tile {
+func NewTile(z, x, y uint, buffer float64, srid uint64) *Tile {
 	return &Tile{
 		z:      z,
 		x:      x,
@@ -14,11 +14,11 @@ func NewTile(z, x, y uint64, buffer float64, srid uint64) *Tile {
 
 type Tile struct {
 	// zoom
-	z uint64
+	z uint
 	// column
-	x uint64
+	x uint
 	// row
-	y uint64
+	y uint
 	// buffer will add a buffer to the tile bounds. this buffer is expected to use the same units as the SRID
 	// of the projected tile (i.e. WebMercator = pixels, 3395 = meters)
 	Buffer float64
@@ -26,7 +26,7 @@ type Tile struct {
 	SRID uint64
 }
 
-func (t *Tile) ZXY() (uint64, uint64, uint64) {
+func (t *Tile) ZXY() (uint, uint, uint) {
 	return t.z, t.x, t.y
 }
